fix code

+++ Lesson_1/task_2.rb
@@ -4,11 +4,11 @@
 # и возвращать его площадь.
 
 puts "Введите основание треугольника"
-base = gets.chomp.to_i
+base = gets.chomp.to_f
 
 puts "Введите высоту треугольника"
-height = gets.chomp.to_i
+height = gets.chomp.to_f
 
-square = 0.5*base*height
+square = 0.5 * base * height
 
 puts "Площадь треугольника равна #{square}"
